WIP refactoring internals
