+++ crates/fta/src/config/mod.rs
@@ -1,9 +1,10 @@
+mod tests;
+
 use crate::structs::FtaConfig;
 use std::fs::File;
 use std::io::Read;
 use std::path::Path;
 
-#[allow(dead_code)]
 pub fn read_config(config_path: &str) -> FtaConfig {
     let default_config = FtaConfig {
         extensions: Some(vec![
